like button moved next to name resource detailscard
